requirements.txt
selenium==4.0.0
webdriver-manager==3.8.1
appium-Python-Client==2.1.0
behave==1.2.6
pytest==6.2.5  # For running tests if you want a test runner